Remove unnecessary module.
